src/app/dashboard/shared/components/profile-form/profile-form.component.spec.ts
// import {
//   beforeEach,
//   beforeEachProviders,
//   describe,
//   expect,
//   it,
//   inject,
// } from '@angular/core/testing';
// import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
// import { Component } from '@angular/core';
// import { By } from '@angular/platform-browser';
// import { ProfileFormComponent } from './profile-form.component';

// describe('Component: ProfileForm', () => {
//   let builder: TestComponentBuilder;

//   beforeEachProviders(() => [ProfileFormComponent]);
//   beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
//     builder = tcb;
//   }));

//   it('should inject the component', inject([ProfileFormComponent],
//       (component: ProfileFormComponent) => {
//     expect(component).toBeTruthy();
//   }));

//   it('should create the component', inject([], () => {
//     return builder.createAsync(ProfileFormComponentTestController)
//       .then((fixture: ComponentFixture<any>) => {
//         let query = fixture.debugElement.query(By.directive(ProfileFormComponent));
//         expect(query).toBeTruthy();
//         expect(query.componentInstance).toBeTruthy();
//       });
//   }));
// });

// @Component({
//   selector: 'test',
//   template: `
//     <app-profile-form></app-profile-form>
//   `,
//   directives: [ProfileFormComponent]
// })
// class ProfileFormComponentTestController {
// }

